+++ src/Provider/Provider.js
@@ -12,13 +12,13 @@ const Keyset = require('../Utils/Keyset')
 
 const GradeService = require('./Services/Grade')
 
-const url = require('url')
+const url = require('fast-url-parser')
 const _path = require('path')
 const jwt = require('jsonwebtoken')
 const winston = require('winston')
 const validUrl = require('valid-url')
 const crypto = require('crypto')
-const { getDomain } = require('tldjs')
+const tldparser = require('tld-extract')
 
 const provAuthDebug = require('debug')('provider:auth')
 const provMainDebug = require('debug')('provider:main')
@@ -641,7 +641,7 @@ class Provider {
       provMainDebug('Setting up path cookie for this resource with path: ' + path)
       const cookieOptions = JSON.parse(JSON.stringify(this.#cookieOptions))
       if (externalRequest) {
-        const domain = getDomain(externalRequest)
+        const domain = tldparser(externalRequest).domain
         cookieOptions.domain = '.' + domain
         provMainDebug('External request found for domain: .' + domain)
       }
@@ -662,8 +662,28 @@ class Provider {
         res.clearCookie(code + this.#appUrl, this.#cookieOptions)
       }
     }
+
+    // Formatting path with queries
+    const pathParts = url.parse(path)
+    const params = new URLSearchParams(pathParts.search)
+    const queries = {}
+    for (const [key, value] of params) { queries[key] = value }
+
+    const formattedPath = url.format({
+      protocol: pathParts.protocol,
+      hostname: pathParts.hostname,
+      pathname: pathParts.pathname,
+      port: pathParts.port,
+      auth: pathParts.auth,
+      query: {
+        ...queries,
+        ltik: token
+      }
+    })
+
+    // Sets allow credentials header and redirects to path with queries
     res.header('Access-Control-Allow-Credentials', 'true')
-    return res.redirect(path + '?ltik=' + token)
+    return res.redirect(formattedPath)
   }
 }